Update array data type for testing purposes

+++ averagecalculator.c
@@ -16,7 +16,7 @@ Remember to handle input validation to ensure that the user enters valid integer
 float calculateAverage(int numbers[], int size);
 
 int main(void) {
-    float arrayForAverages[maxArraySize]; // array of empty values maximum size is 5
+    int arrayForAverages[maxArraySize]; // array of empty values maximum size is 5
     printf("Enter the values for the array (up to a maximum of 5 values:\n");
 
     // h/t King C Programming book for the idiom on reading data into arrays
